Move AdrUdrProduct class into its own files.
